3D MCS: changes based on code review

+++ Code/GraphMol/FMCS/FMCS.h
@@ -44,7 +44,7 @@ struct RDKIT_FMCS_EXPORT MCSAtomCompareParameters {
   bool RingMatchesRingOnly = false;
   bool CompleteRingsOnly = false;
   bool MatchIsotope = false;
-  float MaxDistance = -1.0;
+  double MaxDistance = -1.0;
 };
 
 struct RDKIT_FMCS_EXPORT MCSBondCompareParameters {
